requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0

# PPT generation
python-pptx>=0.6.21
Pillow>=10.0.0

# Image processing and segmentation
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0

# Background removal
onnxruntime>=1.15.0

# Vector graphics conversion
vtracer>=0.3.0
cairosvg>=2.7.0

# MinerU client
mineru-vl-utils>=0.1.0

